Made CurrentMobArea its own class
